+++ gears/tags.py
@@ -216,6 +216,7 @@ ST_JEWELRY = "JEWELRY"
 ST_ANTIQUE = "ANTIQUE"
 ST_CONTRABAND = "CONTRABAND"
 ST_TOOL = "TOOL"
+ST_SURVIVAL = "SURVIVAL"
 
 # Inventory Slots- note that these don't work exactly like slots in most RPGs.
 # Multiple items can be equipped in the same place as long as none of them have the same slot.